chore: move Image Upload constant vars and update imports
