+++ app/ai/components/ai-chat-form.tsx
@@ -11,8 +11,7 @@ import { Button } from "@/components/ui/button";
 import { useAIThemeGeneration } from "@/hooks/use-ai-theme-generation";
 import { useImageUpload } from "@/hooks/use-image-upload";
 import { useDocumentDragAndDropIntent } from "@/hooks/useDocumentDragAndDropIntent";
-import { MAX_IMAGE_FILE_SIZE, MAX_IMAGE_FILES } from "@/lib/ai/ai-theme-generator";
-import { AI_PROMPT_CHARACTER_LIMIT } from "@/lib/constants";
+import { AI_PROMPT_CHARACTER_LIMIT, MAX_IMAGE_FILE_SIZE, MAX_IMAGE_FILES } from "@/lib/constants";
 import { cn } from "@/lib/utils";
 import { AIPromptData } from "@/types/ai";
 import { convertFileArrayToFileList } from "@/utils/file-upload";
