+++ lib/constants.ts
@@ -2,4 +2,7 @@ export const AI_PROMPT_CHARACTER_LIMIT = 500;
 
 export const DEBOUNCE_DELAY = 50;
 
-export const AI_REQUEST_FREE_TIER_LIMIT = 5;
+export const AI_REQUEST_FREE_TIER_LIMIT = process.env.NODE_ENV === "development" ? Infinity : 5;
+
+export const MAX_IMAGE_FILES = 3;
+export const MAX_IMAGE_FILE_SIZE = 5 * 1024 * 1024; // 5MB
